Port to mojo of 2025-11-05
- Changed `ImmutableOrigin` to `ImmutOrigin` in multiple array files for
consistency.
- Updated `unsafe_get` method in `StringArray` to use `origin_of(self)`
instead of `__origin_of(self)`.
- Modified test files to import `TestSuite` for running tests.
- Updated `pixi.toml` to run tests using a shell script instead of a
direct command.

+++ firebolt/io/formatter.mojo
@@ -59,7 +59,7 @@ struct Formatter:
                 break
 
             if value.is_valid(i):
-                writer.write(value.unsafe_get(i))
+                writer.write(value.unsafe_get(UInt(i)))
             else:
                 writer.write("NULL")
         writer.write("])")
